remove the timeout mechanism as the default parallel level is 1
The previous implementation has a timeout mechanism where a request will
stop polling BrowserStack if a job takes longer than expected to finish.
By default, that limit was set to one minute.

However, even if we no longer poll the endpoint, BrowserStack still
consider the job to be running so when we try to send in another
request, BrowserStack will send back `422` instead of `200`

        {"message":"Parallel limit reached","running_sessions":1}

So here we use a simpler `while True` loop to wait even more patiently.